Helper can skip properties itself (make_skipfunc)
This enables sections to specify a whitelist of properties they are
interested in.

By default, the Helper class skips the i18n_domain property.
It is now enabled to use a whitelist *or* a blacklist of property names;
unless given a whitelist, 'i18n_domain' will still be skipped.

A method _skip_this is generated by the factory function make_skipfunc.